Refactor how the AUG curve is calculated
